+++ xrpl/models/requests/generic_request.py
@@ -5,6 +5,8 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Any, Dict, Type, Union, cast
 
+from typing_extensions import Self
+
 from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.requests.request import Request, RequestMethod
 from xrpl.models.utils import KW_ONLY_DATACLASS, require_kwargs_on_init
@@ -26,7 +28,7 @@ class GenericRequest(Request):
     :meta hide-value:
     """
 
-    def __init__(self: GenericRequest, **kwargs: Any) -> None:
+    def __init__(self: Self, **kwargs: Any) -> None:
         """
         Initializes a GenericRequest.
 
@@ -43,7 +45,7 @@ class GenericRequest(Request):
             object.__setattr__(self, key, value)
 
     @classmethod
-    def from_dict(cls: Type[GenericRequest], value: Dict[str, Any]) -> GenericRequest:
+    def from_dict(cls: Type[Self], value: Dict[str, Any]) -> Self:
         """
         Construct a new GenericRequest from a dictionary of parameters. Also converts
         from JSON and WS formatting.
@@ -75,7 +77,7 @@ class GenericRequest(Request):
 
         return cls(**value)
 
-    def to_dict(self: GenericRequest) -> Dict[str, Any]:
+    def to_dict(self: Self) -> Dict[str, Any]:
         """
         Returns the dictionary representation of a GenericRequest.