+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
-- Add support for the DeliverMax field in Payment transactions
 
 ### Added
+- Add support for the DeliverMax field in Payment transactions
 - Use rippled API v2 as default in requests
+
+### Fixed
+- Allow empty strings for the purpose of removing fields in DIDSet transaction
 
 ## [2.6.0] - 2024-06-03